Moved thread class and changed comments
