+++ frontstage/common/thread_wrapper.py
@@ -1,0 +1,11 @@
+import threading
+
+
+class ThreadWrapper(threading.Thread):
+    def __init__(self, function, *args):
+        threading.Thread.__init__(self)
+        self.function = function
+        self.args = args
+
+    def run(self):
+        self.function(*self.args)
